# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AppService:
    properties:
      branch:
        type: string
      buildCommand:
        type: string
      buildEngine:
        type: string
      containerId:
        type: string
      createdAt:
        type: string
      dockerfilePath:
        type: string
      domain:
        type: string
      environmentId:
        type: string
      healthCheckPath:
        type: string
      id:
        type: string
      internalPort:
        type: integer
      name:
        type: string
      projectId:
        type: string
      repositoryUrl:
        type: string
      rootDirectory:
        type: string
      startCommand:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  models.BackupConfig:
    properties:
      createdAt:
        type: string
      databaseId:
        type: string
      id:
        type: string
      name:
        type: string
      projectId:
        type: string
      retentionDays:
        type: integer
      s3DestinationId:
        type: string
      schedule:
        type: string
      status:
        type: string
      storageId:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Vessl API Documentation
  title: Vessl API
  version: "1.0"
paths:
  /api/apps/{id}:
    delete:
      consumes:
      - application/json
      description: Delete App Service
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete App Service
      tags:
      - AppServices
    get:
      consumes:
      - application/json
      description: Get App Service
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get App Service
      tags:
      - AppServices
    put:
      consumes:
      - application/json
      description: Update App Service
      parameters:
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AppService'
      produces:
      - application/json
      responses: {}
      summary: Update App Service
      tags:
      - AppServices
...
  /api/backups:
    get:
      consumes:
      - application/json
      description: List Backups
      produces:
      - application/json
      responses: {}
      summary: List Backups
      tags:
      - Backups
    post:
      consumes:
      - application/json
      description: Create Backup
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BackupConfig'
      produces:
      - application/json
      responses: {}
      summary: Create Backup
      tags:
      - Backups
  /api/backups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Backup
      parameters:
      - description: Backup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete Backup
      tags:
      - Backups
    get:
      consumes:
      - application/json
      description: Get Backup
      parameters:
      - description: Backup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get Backup
      tags:
      - Backups
  /api/backups/{id}/trigger:
    post:
      consumes:
      - application/json
      description: Trigger endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Trigger endpoint
      tags:
      - Backups
...
  /api/environments/{id}/apps:
    get:
      consumes:
      - application/json
      description: ListByEnvironment endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListByEnvironment endpoint
      tags:
      - AppServices
    post:
      consumes:
      - application/json
      description: Create endpoint
      parameters:
      - description: Environment ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AppService'
      produces:
      - application/json
      responses: {}
      summary: Create endpoint
      tags:
      - AppServices
...
  /api/projects/{id}/apps:
    get:
      consumes:
      - application/json
      description: ListByProject endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListByProject endpoint
      tags:
      - AppServices
...
  /api/s3-destinations/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteS3Destination endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeleteS3Destination endpoint
      tags:
      - Backups
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"