# internal/handlers/app_services.go
package handlers

import (
	"net/http"

	"github.com/labstack/echo/v4"

	"vessl.dev/vessl/internal/http/middleware"
	"vessl.dev/vessl/internal/models"
	"vessl.dev/vessl/internal/services"
)

type AppHandler struct {
	appService     *services.AppService
	projectService *services.ProjectService
}

func NewAppHandler(s *services.AppService, ps *services.ProjectService) *AppHandler {
	return &AppHandler{appService: s, projectService: ps}
}

func (h *AppHandler) verifyProjectOwnership(c echo.Context, projectID string) error {
	user := middleware.GetUserClaimsFromContext(c.Request().Context())
	if user == nil || user.Role == "admin" {
		return nil
	}
	p, err := h.projectService.GetProject(c.Request().Context(), projectID)
	if err != nil {
		return c.JSON(http.StatusNotFound, map[string]string{"error": "project not found"})
	}
	if p.TeamID != user.UserID {
		return c.JSON(http.StatusForbidden, map[string]string{"error": "access denied"})
	}
	return nil
}

// @Summary Create endpoint
// @Description Create endpoint
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "Environment ID"
// @Param request body models.AppService true "Payload"
// @Router /api/environments/{id}/apps [post]
func (h *AppHandler) Create(c echo.Context) error {
	envID := c.Param("id")
	var req models.AppService
	if err := c.Bind(&req); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"error": "invalid payload"})
	}
	if req.Name == "" {
		return c.JSON(http.StatusBadRequest, map[string]string{"error": "app service name is required"})
	}
	if err := h.verifyProjectOwnership(c, req.ProjectID); err != nil {
		return err
	}
	req.EnvironmentID = envID
	if req.InternalPort == 0 {
		req.InternalPort = 3000
	}
	created, err := h.appService.CreateAppService(c.Request().Context(), &req)
	if err != nil {
		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
	return c.JSON(http.StatusCreated, created)
}

// @Summary ListByEnvironment endpoint
// @Description ListByEnvironment endpoint
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "id"
// @Router /api/environments/{id}/apps [get]
func (h *AppHandler) ListByEnvironment(c echo.Context) error {
	envID := c.Param("id")
	apps, err := h.appService.ListByEnvironment(c.Request().Context(), envID)
	if err != nil {
		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
	user := middleware.GetUserClaimsFromContext(c.Request().Context())
	if user != nil && user.Role != "admin" {
		var filtered []*models.AppService
		for _, app := range apps {
			p, err := h.projectService.GetProject(c.Request().Context(), app.ProjectID)
			if err == nil && p.TeamID == user.UserID {
				filtered = append(filtered, app)
			}
		}
		return c.JSON(http.StatusOK, filtered)
	}
	return c.JSON(http.StatusOK, apps)
}

// @Summary ListByProject endpoint
// @Description ListByProject endpoint
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "id"
// @Router /api/projects/{id}/apps [get]
func (h *AppHandler) ListByProject(c echo.Context) error {
	projectID := c.Param("id")
	if err := h.verifyProjectOwnership(c, projectID); err != nil {
		return err
	}
	apps, err := h.appService.ListByProject(c.Request().Context(), projectID)
	if err != nil {
		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
	return c.JSON(http.StatusOK, apps)
}

// @Summary Get App Service
// @Description Get App Service
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "App ID"
// @Router /api/apps/{id} [get]
func (h *AppHandler) Get(c echo.Context) error {
	id := c.Param("id")
	svc, err := h.appService.GetAppService(c.Request().Context(), id)
	if err != nil || svc == nil {
		return c.JSON(http.StatusNotFound, map[string]string{"error": "app service not found"})
	}
	if err := h.verifyProjectOwnership(c, svc.ProjectID); err != nil {
		return err
	}
	return c.JSON(http.StatusOK, svc)
}

// @Summary Update App Service
// @Description Update App Service
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "App ID"
// @Param request body models.AppService true "Payload"
// @Router /api/apps/{id} [put]
func (h *AppHandler) Update(c echo.Context) error {
	id := c.Param("id")
	existing, err := h.appService.GetAppService(c.Request().Context(), id)
	if err != nil || existing == nil {
		return c.JSON(http.StatusNotFound, map[string]string{"error": "app service not found"})
	}
	if err := h.verifyProjectOwnership(c, existing.ProjectID); err != nil {
		return err
	}
	var req models.AppService
	if err := c.Bind(&req); err != nil {
		return c.JSON(http.StatusBadRequest, map[string]string{"error": "invalid payload"})
	}
	existing.Name = req.Name
	existing.RepositoryURL = req.RepositoryURL
	existing.Branch = req.Branch
	existing.RootDirectory = req.RootDirectory
	existing.BuildCommand = req.BuildCommand
	existing.StartCommand = req.StartCommand
	existing.DockerfilePath = req.DockerfilePath
	existing.BuildEngine = req.BuildEngine
	existing.InternalPort = req.InternalPort
	existing.Domain = req.Domain
	existing.HealthCheckPath = req.HealthCheckPath
	existing.ContainerID = req.ContainerID
	existing.Status = req.Status
	if err := h.appService.UpdateAppService(c.Request().Context(), existing); err != nil {
		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
	return c.JSON(http.StatusOK, existing)
}

// @Summary Delete App Service
// @Description Delete App Service
// @Tags AppServices
// @Accept json
// @Produce json
// @Param id path string true "App ID"
// @Router /api/apps/{id} [delete]
func (h *AppHandler) Delete(c echo.Context) error {
	id := c.Param("id")
	existing, err := h.appService.GetAppService(c.Request().Context(), id)
	if err != nil || existing == nil {
		return c.JSON(http.StatusNotFound, map[string]string{"error": "app service not found"})
	}
	if err := h.verifyProjectOwnership(c, existing.ProjectID); err != nil {
		return err
	}
	if err := h.appService.DeleteAppService(c.Request().Context(), id); err != nil {
		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
	return c.NoContent(http.StatusNoContent)
}